Fast_ML_API/requirements.txt
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
pyarrow==14.0.0
pymongo==4.6.1
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
notebook==7.0.6
tqdm==4.66.1